+++ packages/core/src/index.ts
@@ -31,3 +31,4 @@ export * from "./recipe-store.js";
 export * from "./custom-provider.js";
 export * from "./script-runner.js";
 export * from "./global-switch.js";
+export * from "./doctor.js";